Reduce duplication in test setup
