+++ codespan-reporting/Cargo.toml
@@ -21,6 +21,7 @@ unicode-width = "0.1"
 
 [dev-dependencies]
 insta = "0.8"
+lazy_static = "1.3"
 structopt = "0.2"
 unindent = "0.1"
